+++ src/containers/Form/CapacityRules/__tests__/index.test.js
@@ -1,0 +1,52 @@
+/*
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
+ */
+
+import React from 'react';
+import { createStore } from 'redux';
+import { shallow } from 'enzyme';
+import { getCurrentForm } from '../../../../redux/modules/form/selectors';
+import CapacityRulesForm, { mapStateToProps } from '../';
+import { getSelectedEntityId, isCreating, userHasUpdatePermission } from '../../../../redux/modules/entities/selectors';
+import { selectFormInitialValues, formSubmission, createFormName } from '../../../../redux/modules/form/selectors';
+
+jest.mock('../../../../redux/modules/entities/selectors');
+jest.mock('../../../../redux/modules/form/selectors');
+getCurrentForm.mockImplementation(() => 'gets form from state');
+getSelectedEntityId.mockImplementation(() => 'mockId');
+isCreating.mockImplementation(() => true);
+userHasUpdatePermission.mockImplementation(() => true);
+selectFormInitialValues.mockImplementation(() => ({ active: true }));
+
+describe('CapacityRules Renders', () => {
+  it('renders', () => {
+    const store = createStore((state) => state);
+    expect(shallow(<CapacityRulesForm store={store}>Child</CapacityRulesForm>)).toMatchSnapshot();
+  });
+});
+
+describe('Maps state to props only using selectors', () => {
+  it('validates object created from mapStateToProps', () => {
+    expect(mapStateToProps()).toMatchSnapshot();
+  });
+});
+
+describe('createFormName', () => {
+  it('returns proper values', () => {
+    expect(createFormName()).toMatchSnapshot();
+  });
+});
+
+describe('formSubmission', () => {
+  const values = {
+    id: 'mockId',
+    name: 'mockName',
+    description: 'mockDescription',
+    type: 'mockType',
+  };
+  const dispatch = (action) => action;
+  const props = { dirty: true };
+  it('returns proper values', () => {
+    expect(formSubmission(values, dispatch, props)).toMatchSnapshot();
+  });
+});