+++ src/containers/Form/IdentityProviders/layout.js
@@ -47,54 +47,52 @@ export default class IdentityProvidersForm extends Component {
   };
 
   render() {
-    const {
-      props: {
-        intl: { formatMessage },
-      },
-    } = this;
+    const { props: { intl: { formatMessage } } } = this;
 
     const initValues = this.props.initialValues.toJS();
-    const idpTypes = (!initValues.id) ?
-      this.props.identityProviderTypes :
-      this.props.identityProviderTypes.filter(
-        idpType => idpType.value !== 'sharedIdentityProviderLinkId');
-    idpTypes.map((type) => {
-      switch(type.value) {
+    const idpTypes = !initValues.id
+      ? this.props.identityProviderTypes
+      : this.props.identityProviderTypes.filter((idpType) => idpType.value !== 'sharedIdentityProviderLinkId');
+    idpTypes.forEach((type) => {
+      switch (type.value) {
         case 'url': {
           type.label = formatMessage({
             id: 'identityProviders.details.selectedIdpConfigInfoType.url',
-            defaultMessage: 'URL'
+            defaultMessage: 'URL',
           });
           break;
         }
         case 'xml': {
           type.label = formatMessage({
             id: 'identityProviders.details.selectedIdpConfigInfoType.xml',
-            defaultMessage: 'XML File'
+            defaultMessage: 'XML File',
           });
           break;
         }
         case 'xmlDirectInput': {
           type.label = formatMessage({
             id: 'identityProviders.details.selectedIdpConfigInfoType.xmlDirectInput',
-            defaultMessage: 'XML Direct Input'
+            defaultMessage: 'XML Direct Input',
           });
           break;
         }
         case 'sharedIdentityProviderLinkId': {
           type.label = formatMessage({
             id: 'identityProviders.details.selectedIdpConfigInfoType.sharedIdentityProviderAccessCode',
-            defaultMessage: 'Shared Access Code'
+            defaultMessage: 'Shared Access Code',
           });
           break;
         }
+        default:
+          return;
       }
     });
     return (
       <form onSubmit={this.props.handleSubmit} key={this.props.key}>
-        <DetailHeader text={`${formatMessage({
+        <DetailHeader
+          text={`${formatMessage({
             id: 'identityProviders.details.details',
-            defaultMessage: 'Details'
+            defaultMessage: 'Details',
           })}`}
         />
         <InputField
@@ -102,7 +100,7 @@ export default class IdentityProvidersForm extends Component {
           name="name"
           label={`${formatMessage({
             id: 'identityProviders.details.name',
-            defaultMessage: 'Name'
+            defaultMessage: 'Name',
           })} *`}
           data-automation="nameInput"
           componentType="input"
@@ -115,7 +113,7 @@ export default class IdentityProvidersForm extends Component {
           name="description"
           label={`${formatMessage({
             id: 'identityProviders.details.description',
-            defaultMessage: 'Description'
+            defaultMessage: 'Description',
           })}`}
           data-automation="descriptionInput"
           componentType="textarea"
@@ -126,27 +124,31 @@ export default class IdentityProvidersForm extends Component {
           name="identityProviderType"
           label={`${formatMessage({
             id: 'identityProviders.details.type',
-            defaultMessage: 'Type'
+            defaultMessage: 'Type',
           })} *`}
           placeholder={`${formatMessage({
             id: 'identityProviders.details.idpConfigInfoTypes.choose',
-            defaultMessage: 'Select type...'
+            defaultMessage: 'Select type...',
           })} *`}
           data-automation="identityProviderTypeList"
           disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
-          options={[{
-            label: formatMessage({
-              id: 'identityProviders.details.idpConfigInfoTypes.choose',
-              defaultMessage: 'Select type...'
-            }),
-            value: ''
-          }, ...idpTypes]}
+          options={[
+            {
+              label: formatMessage({
+                id: 'identityProviders.details.idpConfigInfoTypes.choose',
+                defaultMessage: 'Select type...',
+              }),
+              value: '',
+            },
+            ...idpTypes,
+          ]}
           required
         />
 
-        <DetailHeader text={`${formatMessage({
+        <DetailHeader
+          text={`${formatMessage({
             id: 'identityProviders.details.configuration',
-            defaultMessage: 'Configuration'
+            defaultMessage: 'Configuration',
           })}`}
         />
         {(this.props.idpType === 'url' || this.props.idpType === 'xml') && (
@@ -155,7 +157,7 @@ export default class IdentityProvidersForm extends Component {
             name="emailMapping"
             label={`${formatMessage({
               id: 'identityProviders.details.emailMapping',
-              defaultMessage: 'Email Mapping'
+              defaultMessage: 'Email Mapping',
             })} *`}
             data-automation="emailMappingInput"
             componentType="input"
@@ -165,13 +167,13 @@ export default class IdentityProvidersForm extends Component {
           />
         )}
         {((!initValues.id && this.props.idpType === 'url') ||
-        (this.props.idpType === 'url' && this.state.showMetadataUrlInput)) && (
+          (this.props.idpType === 'url' && this.state.showMetadataUrlInput)) && (
           <InputField
             className="frm-identityproviders-metadataUrl"
             name="metadataUrl"
             label={`${formatMessage({
               id: 'identityProviders.details.metadataUrl',
-              defaultMessage: 'Metadata URL'
+              defaultMessage: 'Metadata URL',
             })} *`}
             data-automation="metadataUrlInput"
             componentType="input"
@@ -180,67 +182,70 @@ export default class IdentityProvidersForm extends Component {
             required
           />
         )}
-        {(initValues.id && this.props.idpType === 'url' &&
-          !this.state.showMetadataUrlInput) && (
+        {initValues.id &&
+          this.props.idpType === 'url' &&
+          !this.state.showMetadataUrlInput && (
             <Detail
               name="metadataUrl"
               label={`${formatMessage({
                 id: 'identityProviders.details.metadataUrl',
-                defaultMessage: 'Metadata URL'
+                defaultMessage: 'Metadata URL',
               })} *`}
               value={
                 initValues.metadataUrl
                   ? initValues.metadataUrl
                   : formatMessage({
                       id: 'identityProviders.details.metadataUrl.noUrl',
-                      defaultMessage: '(No URL to XML Config Set)'
+                      defaultMessage: '(No URL to XML Config Set)',
                     })
               }
             />
           )}
-        {initValues.id && this.props.idpType === 'url' &&
+        {initValues.id &&
+          this.props.idpType === 'url' &&
           initValues.metadataUrl &&
           !this.state.showMetadataUrlInput && (
             <A onClick={() => this.setState({ showMetadataUrlInput: true })}>
               {`${formatMessage({
                 id: 'identityProviders.details.editXmlUrl',
-                defaultMessage: 'edit'
+                defaultMessage: 'edit',
               })}`}
             </A>
           )}
-        {initValues.id && this.props.idpType === 'url' &&
+        {initValues.id &&
+          this.props.idpType === 'url' &&
           !initValues.metadataUrl &&
           !this.state.showMetadataUrlInput && (
             <A onClick={() => this.setState({ showMetadataUrlInput: true })}>
-            {`${formatMessage({
-              id: 'identityProviders.details.addXmlUrl',
-              defaultMessage: 'add url'
-            })}`}
+              {`${formatMessage({
+                id: 'identityProviders.details.addXmlUrl',
+                defaultMessage: 'add url',
+              })}`}
             </A>
           )}
 
-          {this.props.idpType === 'xml' && (
-            <>
-              <FileUploadField
-                uploadFile={()=> ""}
-                fileType="text/xml"
-                acceptedFileType="text/xml"
-                label={`${formatMessage({
-                  id: 'identityProviders.details.uploadXmlConfig',
-                  defaultMessage: 'Upload XML Config'
-                })} *`}
-                disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
-                id="newMetadataFile"
-                name="newMetadataFile"
-                maxFileSize={1000000}
-                toastError={`${formatMessage({
-                  id: 'identityProviders.details.uploadXmlConfig.error',
-                  defaultMessage: 'File must be an XML under 10MB'
-                })} *`}
-                required
-              />
-            </>
-          )}
+        {this.props.idpType === 'xml' && (
+          <>
+            <FileUploadField
+              uploadFile={() => ''}
+              fileType="text/xml"
+              acceptedFileType="text/xml"
+              label={`${formatMessage({
+                id: 'identityProviders.details.uploadXmlConfig',
+                defaultMessage: 'Upload XML Config',
+              })} *`}
+              disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
+              id="newMetadataFile"
+              name="newMetadataFile"
+              maxFileSize={1000000}
+              toastError={`${formatMessage({
+                id: 'identityProviders.details.uploadXmlConfig.error',
+                defaultMessage: 'File must be an XML under 10MB',
+              })} *`}
+              required
+            />
+          </>
+        )}
 
         {this.props.idpType === 'xml' &&
           initValues.metadataFile && (
@@ -250,7 +255,7 @@ export default class IdentityProvidersForm extends Component {
                 onClick={(e) => this.downloadFile(e, initValues)}
                 value={`${formatMessage({
                   id: 'identityProviders.details.downloadXmlConfig',
-                  defaultMessage: 'Download Current Config'
+                  defaultMessage: 'Download Current Config',
                 })}`}
                 disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
                 id="dtpanel-identity-providers-download-xml"
@@ -259,19 +264,19 @@ export default class IdentityProvidersForm extends Component {
           )}
 
         {this.props.idpType === 'xmlDirectInput' && (
-            <InputField
-              className="frm-identityproviders-metadataFile"
-              name="metadataFile"
-              label={`${formatMessage({
-                id: 'identityProviders.details.enterXml',
-                defaultMessage: 'Enter XML Markup Here'
-              })} *`}
-              data-automation="metadataFileInput"
-              componentType="textarea"
-              inputType="text"
-              disabled={this.state.disableXmlInputEdit && initValues.metadataFile ? true : false}
-              required
-            />
+          <InputField
+            className="frm-identityproviders-metadataFile"
+            name="metadataFile"
+            label={`${formatMessage({
+              id: 'identityProviders.details.enterXml',
+              defaultMessage: 'Enter XML Markup Here',
+            })} *`}
+            data-automation="metadataFileInput"
+            componentType="textarea"
+            inputType="text"
+            disabled={this.state.disableXmlInputEdit && initValues.metadataFile ? true : false}
+            required
+          />
         )}
         {this.props.idpType === 'xmlDirectInput' &&
           this.state.disableXmlInputEdit &&
@@ -279,38 +284,41 @@ export default class IdentityProvidersForm extends Component {
             <A onClick={() => this.setState({ disableXmlInputEdit: false })}>
               {`${formatMessage({
                 id: 'identityProviders.details.editXmlMarkup',
-                defaultMessage: 'edit xml markup'
+                defaultMessage: 'edit xml markup',
               })}`}
             </A>
           )}
         {this.props.idpType === 'xmlDirectInput' &&
-          !this.state.disableXmlInputEdit && <A onClick={() => this.setState({ disableXmlInputEdit: true })}>
-            {`${formatMessage({
-              id: 'identityProviders.details.cancelEditXmlMarkup',
-              defaultMessage: 'cancel'
-            })}`}
-          </A>}
-        {(!initValues.id && this.props.idpType === 'sharedIdentityProviderLinkId') && (
-          <InputField
-            className="frm-identityproviders-identityProvider"
-            name="identityProvider"
-            label={`${formatMessage({
-              id: 'identityProviders.details.sharedAccessCode',
-              defaultMessage: 'Access Code'
-            })} *`}
-            data-automation="identityProviderInput"
-            componentType="input"
-            inputType="text"
-            disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
-            required
-          />
-        )}
+          !this.state.disableXmlInputEdit && (
+            <A onClick={() => this.setState({ disableXmlInputEdit: true })}>
+              {`${formatMessage({
+                id: 'identityProviders.details.cancelEditXmlMarkup',
+                defaultMessage: 'cancel',
+              })}`}
+            </A>
+          )}
+        {!initValues.id &&
+          this.props.idpType === 'sharedIdentityProviderLinkId' && (
+            <InputField
+              className="frm-identityproviders-identityProvider"
+              name="identityProvider"
+              label={`${formatMessage({
+                id: 'identityProviders.details.sharedAccessCode',
+                defaultMessage: 'Access Code',
+              })} *`}
+              data-automation="identityProviderInput"
+              componentType="input"
+              inputType="text"
+              disabled={this.props.isSaving || !this.props.userHasUpdatePermission}
+              required
+            />
+          )}
         {initValues.id && (
           <Detail
             name="identityProvider"
             label={`${formatMessage({
               id: 'identityProviders.details.sharedIdentityProviderId',
-              defaultMessage: 'Shared Identity Provider ID'
+              defaultMessage: 'Shared Identity Provider ID',
             })} *`}
             value={initValues.identityProvider}
           />